Make Numeric json repr more useful
