+++ core/src/numeric.rs
@@ -10,7 +10,7 @@ use crate::bigrat::BigRat;
 
 /// Number type.
 #[derive(Clone, PartialEq, Debug, Serialize)]
-#[serde(untagged)]
+#[serde(into = "NumericParts")]
 pub enum Numeric {
     /// Arbitrary-precision rational fraction.
     Rational(BigRat),
@@ -26,6 +26,24 @@ pub enum Numeric {
 enum Parity {
     Rational(BigRat, BigRat),
     Float(f64, f64),
+}
+
+/// Used when converting to string representation to choose desired
+/// output mode.
+#[derive(Debug, Clone, Copy, PartialEq, Serialize)]
+pub enum Digits {
+    Default,
+    FullInt,
+    Digits(u64),
+}
+
+#[derive(Debug, Clone, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct NumericParts {
+    numer: String,
+    denom: String,
+    exact_value: Option<String>,
+    approx_value: Option<String>,
 }
 
 impl Numeric {
@@ -127,6 +145,105 @@ impl Numeric {
     pub fn to_f64(&self) -> f64 {
         self.into()
     }
+
+    pub fn to_string(&self, base: u8, digits: Digits) -> (bool, String) {
+        use std::char::from_digit;
+        let sign = *self < Numeric::zero();
+        let rational = self.abs();
+        let (num, den) = rational.to_rational();
+        let rational = match rational {
+            Numeric::Rational(rational) => rational,
+            Numeric::Float(f) => BigRat::from(f),
+        };
+        let intdigits = (&num / &den).size_in_base(base) as u32;
+        let mut buf = String::new();
+        if sign {
+            buf.push('-');
+        }
+        let zero = BigRat::zero();
+        let one = BigInt::one();
+        let ten = BigInt::from(base as u64);
+        let ten_rational = BigRat::ratio(&ten, &one);
+        let mut cursor = &rational / &BigRat::ratio(&ten.pow(intdigits), &one);
+        let mut n = 0;
+        let mut only_zeros = true;
+        let mut zeros = 0;
+        let mut placed_decimal = false;
+        loop {
+            let exact = cursor == zero;
+            let use_sci = if digits != Digits::Default
+                || den == one && (base == 2 || base == 8 || base == 16 || base == 32)
+            {
+                false
+            } else {
+                intdigits + zeros > 9 * 10 / base as u32
+            };
+            let placed_ints = n >= intdigits;
+            let ndigits = match digits {
+                Digits::Default | Digits::FullInt => 6,
+                Digits::Digits(n) => intdigits as i32 + n as i32,
+            };
+            let bail = (exact && (placed_ints || use_sci))
+                || (n as i32 - zeros as i32 > ndigits && use_sci)
+                || n as i32 - zeros as i32 > ::std::cmp::max(intdigits as i32, ndigits);
+            if bail && use_sci {
+                // scientific notation
+                let off = if n < intdigits { 0 } else { zeros };
+                buf = buf[off as usize + placed_decimal as usize + sign as usize..].to_owned();
+                buf.insert(1, '.');
+                if buf.len() == 2 {
+                    buf.insert(2, '0');
+                }
+                if sign {
+                    buf.insert(0, '-');
+                }
+                buf.push_str(&*format!("e{}", intdigits as i32 - zeros as i32 - 1));
+                return (exact, buf);
+            }
+            if bail {
+                return (exact, buf);
+            }
+            if n == intdigits {
+                buf.push('.');
+                placed_decimal = true;
+            }
+            let digit = &(&(&cursor.numer() * &ten) / &cursor.denom()) % &ten;
+            let v: Option<i64> = digit.as_int();
+            let v = v.unwrap();
+            if v != 0 {
+                only_zeros = false
+            } else if only_zeros {
+                zeros += 1;
+            }
+            if !(v == 0 && only_zeros && n < intdigits - 1) {
+                buf.push(from_digit(v as u32, base as u32).unwrap());
+            }
+            cursor = &cursor * &ten_rational;
+            cursor = &cursor - &BigRat::ratio(&digit, &one);
+            n += 1;
+        }
+    }
+
+    pub fn string_repr(&self, base: u8, digits: Digits) -> (Option<String>, Option<String>) {
+        match *self {
+            Numeric::Rational(ref rational) => {
+                let num = rational.numer();
+                let den = rational.denom();
+
+                match self.to_string(base, digits) {
+                    (true, v) => (Some(v), None),
+                    (false, v) => {
+                        if den > BigInt::from(1_000u64) || num > BigInt::from(1_000_000u64) {
+                            (None, Some(v))
+                        } else {
+                            (Some(format!("{}/{}", num, den)), Some(v))
+                        }
+                    }
+                }
+            }
+            Numeric::Float(_f) => (None, Some(self.to_string(base, digits).1)),
+        }
+    }
 }
 
 impl From<BigRat> for Numeric {
@@ -152,6 +269,19 @@ impl<'a> Into<f64> for &'a Numeric {
         match *self {
             Numeric::Rational(ref rational) => rational.as_float(),
             Numeric::Float(f) => f,
+        }
+    }
+}
+
+impl Into<NumericParts> for Numeric {
+    fn into(self) -> NumericParts {
+        let (exact, approx) = self.string_repr(10, Digits::Default);
+        let (num, den) = self.to_rational();
+        NumericParts {
+            numer: num.to_string(),
+            denom: den.to_string(),
+            exact_value: exact,
+            approx_value: approx,
         }
     }
 }
