+++ web/src/components/Numeric.svelte
@@ -1,0 +1,11 @@
+<script lang="typescript">
+  import * as expr from "../util/expr";
+
+  export let number: expr.NumericParts;
+</script>
+
+{#if number.exactValue}
+  {number.exactValue}
+{:else if number.approxValue}
+  {number.approxValue}
+{:else}{number.numer}/{number.denom}{/if}
